Use cloc to measure the lines of code
Couting the lines of code is useful to know how much code is under our responsibility. The lesser the better.

If using a library increase too much this metric, then do not use this library yet. On the contrary, if there is too much boilerplate code and duplications, a library can take the responsibility in our place and reduce the lines of code.

I might try React to check if it decreases the cloc.